shift unit-test folders out from src folder to avoid build construct of test files. added separate unit-tests markdown instructions
